Move devices and directories into file_sys
I've also updated the copyright dates on all these files.
