+++ src/core/file_sys/devices/deci_tty6_device.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: Copyright 2024 shadPS4 Emulator Project
+// SPDX-FileCopyrightText: Copyright 2025 shadPS4 Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
 #pragma once
@@ -7,14 +7,14 @@
 
 namespace Core::Devices {
 
-class URandomDevice final : BaseDevice {
+class DeciTty6Device final : BaseDevice {
     u32 handle;
 
 public:
     static std::shared_ptr<BaseDevice> Create(u32 handle, const char*, int, u16);
-    explicit URandomDevice(u32 handle) : handle(handle) {}
+    explicit DeciTty6Device(u32 handle) : handle(handle) {}
 
-    ~URandomDevice() override = default;
+    ~DeciTty6Device() override = default;
 
     int ioctl(u64 cmd, Common::VaCtx* args) override;
     s64 write(const void* buf, size_t nbytes) override;
