+++ src/core/file_sys/directories/base_directory.h
