+++ src/core/file_sys/directories/normal_directory.h
