+++ src/core/file_sys/directories/pfs_directory.cpp
